+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 ### Added
 - `Machine.env`: Easily get the value of an environment variable
 - Allow specifying a command when spawning a subshell.
-- `ub.bootlog` and `lnx.bootlog` to allow accessing the bootlog (#5)
+- `ub.bootlog` and `lnx.bootlog` to allow accessing the bootlog (ref #5)
+- Add `SSHMachine.ssh_config`: List of additional ssh config options
 
 
 ## [0.6.1] - 2018-11-16